Structural change of code base for data feeder
